refactor: move ErrorMessage component from utils to components
